updated setup script to include some error handling

+++ setup.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-#  Copyright (C) 2015 University of Stuttgart
+#  Copyright (C) 2015, 2016 University of Stuttgart
 #
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -31,16 +31,30 @@ INSTALL_PATH_NVIDIA=`pwd`/${BINARY_FOLDER}/nvidia
 # ============================================================================ #
 # VERSIONS OF REQUIRED LIBRARIES                                               #
 # ============================================================================ #
-PAPI="papi-5.4.0"
+
+PAPI="papi"
+PAPI_VERSION="5.4.0"
+LIKWID="likwid"
+LIKWID_VERSION="3.1.3"
+CURL="curl"
+CURL_VERSION="7.37.0"
+APR="apr"
+APR_VERSION="1.5.2"
+APR_UTIL="apr-util"
+APR_UTIL_VERSION="1.5.4"
 
 # ============================================================================ #
 # DOWNLOAD AND INSTALL PAPI-C                                                  #
 # ============================================================================ #
 
 cd $ROOT
-wget http://icl.cs.utk.edu/projects/papi/downloads/${PAPI}.tar.gz
-tar zxvf ${PAPI}.tar.gz
-cd ${PAPI}/src
+wget http://icl.cs.utk.edu/projects/papi/downloads/${PAPI}-${PAPI_VERSION}.tar.gz
+if [ ! -f ${PAPI}-${PAPI_VERSION}.tar.gz ]; then
+    echo "[ERROR] File not found: " ${PAPI}-${PAPI_VERSION}.tar.gz
+    exit 0;
+fi
+tar zxvf ${PAPI}-${PAPI_VERSION}.tar.gz
+cd ${PAPI}-${PAPI_VERSION}/src
 ./configure --prefix=${INSTALL_PATH_PAPI} --with-components="rapl coretemp infiniband"
 make
 make install all
@@ -50,9 +64,13 @@ make install all
 # ============================================================================ #
 
 cd $ROOT
-wget http://ftp.fau.de/pub/likwid/likwid-3.1.3.tar.gz
-tar zxvf likwid-3.1.3.tar.gz
-cd likwid-3.1.3
+wget http://ftp.fau.de/pub/likwid/${LIKWID}-${LIKWID_VERSION}.tar.gz
+if [ ! -f ${LIKWID}-${LIKWID_VERSION}.tar.gz ]; then
+    echo "[ERROR] File not found: " ${LIKWID}-${LIKWID_VERSION}.tar.gz
+    exit 0;
+fi
+tar zxvf ${LIKWID}-${LIKWID_VERSION}.tar.gz
+cd ${LIKWID}-${LIKWID_VERSION}
 sed -i 's/^SHARED_LIBRARY\s=\sfalse/SHARED_LIBRARY = true/' config.mk
 sed -i -e 's@^PREFIX\s=\s/usr/local@PREFIX = '"$INSTALL_PATH_LIKWID"'@' config.mk
 make
@@ -65,9 +83,13 @@ make install
 # ============================================================================ #
 
 cd $ROOT
-wget http://curl.haxx.se/download/curl-7.37.0.tar.gz
-tar zxvf curl-7.37.0.tar.gz
-cd curl-7.37.0
+wget http://curl.haxx.se/download/${CURL}-${CURL_VERSION}.tar.gz
+if [ ! -f ${CURL}-${CURL_VERSION}.tar.gz ]; then
+    echo "[ERROR] File not found: " ${CURL}-${CURL_VERSION}.tar.gz
+    exit 0;
+fi
+tar zxvf ${CURL}-${CURL_VERSION}.tar.gz
+cd ${CURL}-${CURL_VERSION}
 ./configure --prefix=${INSTALL_PATH_CURL}
 make
 make install
@@ -78,9 +100,13 @@ make install all
 # ============================================================================ #
 
 cd $ROOT
-wget http://mirror.23media.de/apache/apr/apr-1.5.1.tar.gz
-tar zxvf apr-1.5.1.tar.gz
-cd apr-1.5.1
+wget http://www.eu.apache.org/dist/apr/${APR}-${APR_VERSION}.tar.gz
+if [ ! -f ${APR}-${APR_VERSION}.tar.gz ]; then
+    echo "[ERROR] File not found: " ${APR}-${APR_VERSION}.tar.gz
+    exit 0;
+fi
+tar zxvf ${APR}-${APR_VERSION}.tar.gz
+cd ${APR}-${APR_VERSION}
 ./configure --prefix=${INSTALL_PATH_APR}
 make
 make install
@@ -91,9 +117,13 @@ make install all
 # ============================================================================ #
 
 cd $ROOT
-wget http://mirror.23media.de/apache/apr/apr-util-1.5.3.tar.gz
-tar zxvf apr-util-1.5.3.tar.gz
-cd apr-util-1.5.3
+wget http://www.eu.apache.org/dist//apr/apr-util-1.5.4.tar.gz
+if [ ! -f ${APR_UTIL}-${APR_UTIL_VERSION}.tar.gz ]; then
+    echo "[ERROR] File not found: " ${APR_UTIL}-${APR_UTIL_VERSION}.tar.gz
+    exit 0;
+fi
+tar zxvf ${APR_UTIL}-${APR_UTIL_VERSION}.tar.gz
+cd ${APR_UTIL}-${APR_UTIL_VERSION}
 ./configure --prefix=${INSTALL_PATH_APU} --with-apr=${INSTALL_PATH_APR}
 make
 make install
@@ -104,11 +134,17 @@ make install all
 # ============================================================================ #
 
 cd $ROOT
-mkdir nvidia_gdk_download
+mkdir -f nvidia_gdk_download
 cd nvidia_gdk_download
-wget http://developer.download.nvidia.com/compute/cuda/7.5/Prod/gdk/gdk_linux_amd64_352_55_release.run
-chmod +x gdk_linux_amd64_352_55_release.run
-./gdk_linux_amd64_352_55_release.run --silent --installdir=${INSTALL_PATH_NVIDIA}
+NVIDIA_BASE_URL="http://developer.download.nvidia.com"
+NVIDIA_GDK="gdk_linux_amd64_352_55_release.run"
+wget ${NVIDIA_BASE_URL}/compute/cuda/7.5/Prod/gdk/${NVIDIA_GDK}
+if [ ! -f ${NVIDIA_GDK} ]; then
+    echo "[ERROR] File not found: " ${NVIDIA_GDK}
+    exit 0;
+fi
+chmod +x ${NVIDIA_GDK}
+./${NVIDIA_GDK} --silent --installdir=${INSTALL_PATH_NVIDIA}
 
 # ============================================================================ #
 # CLEANING UP                                                                  #
@@ -116,17 +152,16 @@ chmod +x gdk_linux_amd64_352_55_release.run
 
 cd $ROOT
 rm -f *.tar.gz
-rm -rf papi-5.4.0
-rm -rf likwid-3.1.3
-rm -rf apr-1.5.1
-rm -rf apr-util-1.5.3
-rm -rf curl-7.37.0
+rm -rf ${PAPI}-${PAPI_VERSION}
+rm -rf ${LIKWID}-${LIKWID_VERSION}
+rm -rf ${APR}-${APR_VERSION}
+rm -rf ${APR_UTIL}-${APR_UTIL_VERSION}
+rm -rf ${CURL}-${CURL_VERSION}
 rm -rf nvidia_gdk_download
+rm -rf ${NVIDIA_GDK}
 
 # ============================================================================ #
-# COMPILING                                                                    #
+# DONE
 # ============================================================================ #
 
-cd $ROOT
-make
-make install
+echo "\n\nALL DEPENDENCIES WERE INSTALLED SUCCESSFULLY!\n"
